Refactor: Move MailServerDirection into core:mail:mailserver module
