+++ core/mail/mailserver/src/main/kotlin/net/thunderbird/core/mail/mailserver/MailServerDirection.kt
@@ -1,4 +1,4 @@
-package com.fsck.k9.mail
+package net.thunderbird.core.mail.mailserver
 
 enum class MailServerDirection {
     INCOMING,
