+++ legacy/core/src/main/java/com/fsck/k9/LocalKeyStoreManager.kt
@@ -1,10 +1,10 @@
 package com.fsck.k9
 
 import app.k9mail.legacy.account.Account
-import com.fsck.k9.mail.MailServerDirection
 import com.fsck.k9.mail.ssl.LocalKeyStore
 import java.security.cert.CertificateException
 import java.security.cert.X509Certificate
+import net.thunderbird.core.mail.mailserver.MailServerDirection
 
 class LocalKeyStoreManager(
     private val localKeyStore: LocalKeyStore,
