+++ settings.gradle.kts
@@ -221,3 +221,4 @@ check(JavaVersion.current().isCompatibleWith(JavaVersion.VERSION_17)) {
     """.trimIndent()
 }
 include(":core:android:logging")
+include(":core:mail:mailserver")
